Generalize worker thread. Partially implement task queue.
