+++ engine/lib/phx/src/engine/task_queue/mod.rs
@@ -1,6 +1,4 @@
 mod error;
-mod lua_worker;
-mod native_worker;
 mod task_queue;
 mod worker;
 mod worker_data;
@@ -8,8 +6,6 @@ mod worker_thread;
 mod workers;
 
 pub use error::*;
-pub use lua_worker::*;
-pub use native_worker::*;
 pub use task_queue::*;
 pub use worker::*;
 pub use worker_data::*;
