+++ engine/lib/phx/src/engine/task_queue/worker.rs
@@ -1,7 +1,19 @@
+pub type WorkerId = u8;
+
 pub enum Worker {
-    /// Basic worker that replicates an input value into the output
+    /// Example worker that replicates input value into the output
     Echo,
 
     /// Specifies number of engine worker types
     EngineWorkersCount, // !!! SHOULD BE THE LAST ENUM VARIANT !!!
+}
+
+impl Worker {
+    pub fn from_worker_id(id: WorkerId) -> Option<Self> {
+        if id == Self::Echo as WorkerId {
+            Some(Self::Echo)
+        } else {
+            None
+        }
+    }
 }
