+++ engine/lib/phx/src/engine/task_queue/worker_data.rs
@@ -1,3 +1,5 @@
+pub type JobId = usize;
+
 pub enum WorkerInData<T> {
     Ping,
     Data(T),
